+++ src/main/java/com/frejdh/util/common/invocations/ThrowableUtils.java
@@ -1,18 +1,21 @@
-package com.frejdh.util.common.toolbox;
+package com.frejdh.util.common.invocations;
 
 import com.frejdh.util.common.functional.ThrowingSupplier;
 import org.apache.commons.lang3.exception.ExceptionUtils;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
 import static com.frejdh.util.common.toolbox.CommonUtils.sneakyThrow;
 
 public class ThrowableUtils<T> {
 
 	private final List<ThrowableCondition<T>> throwableConditions = new ArrayList<>();
+	private final List<EqualsCondition<T>> equalsToConditions = new ArrayList<>();
 	private final ThrowingSupplier<T> action;
 
 	protected ThrowableUtils(ThrowingSupplier<T> action) {
@@ -43,12 +46,32 @@ public class ThrowableUtils<T> {
 		return throwsException(throwableClasses != null ? Arrays.stream(throwableClasses).collect(Collectors.toList()) : null);
 	}
 
+	// Lower priority than exceptions
+	public EqualsCondition<T> equalsToAny(List<T> equalsToValues) {
+		if (equalsToValues == null) {
+			return null;
+		}
+		EqualsCondition<T> condition = new EqualsCondition<>(this, equalsToValues);
+		equalsToConditions.add(condition);
+		return condition;
+	}
+
+	@SafeVarargs
+	public final EqualsCondition<T> equalsToAny(T... equalsToValue) {
+		return equalsToAny(equalsToValue != null ? Arrays.stream(equalsToValue).collect(Collectors.toList()) : null);
+	}
+
+	public final EqualsCondition<T> equalsToNull() {
+		return equalsToAny(Collections.singletonList((T) null));
+	}
+
 	public T execute() {
+		T retval = null;
 		try {
-			return action.get();
-		} catch (Throwable caughtException) {
+			retval = action.get();
+		} catch (Throwable caughtException) { // Check for ThrowableCondition
 			Throwable rootCause = ExceptionUtils.getRootCause(caughtException);
-			final Throwable e = rootCause != null ? rootCause : caughtException;
+			final Throwable e = (rootCause != null) ? rootCause : caughtException;
 
 			Optional<ThrowableCondition<T>> matchingCondition = throwableConditions.stream()
 					.peek(condition -> condition.throwableClasses = condition.throwableClasses.stream()
@@ -62,53 +85,36 @@ public class ThrowableUtils<T> {
 				ThrowableCondition<T> condition = matchingCondition.get();
 
 				if (condition.hasReturnValue()) {
-					return condition.returnValue;
+					retval = condition.returnValue.get();
 				}
 				else if (condition.hasThrowableValue()) {
 					sneakyThrow(condition.throwableValue);
 				}
 			}
-
-			sneakyThrow(caughtException);
-			return null;
-		}
-	}
-
-
-	/*
-	 * Additional classes
-	 */
-
-	public static class ThrowableCondition<T> {
-		protected ThrowableUtils<T> parent;
-		protected List<Class<? extends Throwable>> throwableClasses;
-		protected boolean hasReturnValue;
-		protected T returnValue;
-		protected Throwable throwableValue;
-
-		public ThrowableCondition(ThrowableUtils<T> parent, List<Class<? extends Throwable>> throwableClasses) {
-			this.parent = parent;
-			this.throwableClasses = throwableClasses;
+			else {
+				sneakyThrow(caughtException);
+			}
 		}
 
-		public ThrowableUtils<T> thenReturn(T returnValue) {
-			this.hasReturnValue = true;
-			this.returnValue = returnValue;
-			return parent;
+		// Check for EqualsCondition
+		if (equalsToConditions.isEmpty()) {
+			return retval;
 		}
 
-		public ThrowableUtils<T> thenThrow(Throwable throwable) {
-			this.throwableValue = throwable;
-			return parent;
-		}
-
-		public boolean hasReturnValue() {
-			return hasReturnValue;
-		}
-
-		public boolean hasThrowableValue() {
-			return throwableValue != null;
-		}
+		final T finalRetval = retval;
+		Optional<EqualsCondition<T>> matchingCondition = equalsToConditions.stream()
+				.peek(condition -> condition.equalsToValue = condition.equalsToValue.stream()
+						.filter(conditionRetval -> {
+							if (finalRetval == null && conditionRetval == null) {
+								return true;
+							}
+							return conditionRetval != null && conditionRetval.equals(finalRetval);
+						})
+						.collect(Collectors.toList())
+				)
+				.filter(condition -> !condition.equalsToValue.isEmpty())
+				.findFirst();
+		return (matchingCondition.isPresent()) ? matchingCondition.get().returnValue.get() : finalRetval;
 	}
 
 }